Use a position controller when using fake hardware.
The position controller mock uses fake position data and then the arm actually moves.
Not sure if this is the right solution, probably the correct solution is an effort controller
that outputs position. Maybe this is what the joint_effort_trajectory_controller provided
by franka is actually doing. (Not sure what it meant by saying joint_trajectory_controller did not support effort...)

- On the other hand, maybe the problem is on the mock hardware interface level. Where the mock effort controllers
should be updating the joint states. Of course then we'd need some simulation of how effort-> position...