Finished refactoring launchfile.
Everything appears to load with fake hardware.
The end-effector is slightly tilted at the wrong angle (I think due to TCP redefinition, so redefining this
needs to be done while checking with the real robot to see if its the fake_hardware position that changes
or the urdf file).

Also need to setup more planning parameters because motion planning fails