Added franka_moveit_config
Franka_moveit_config is the moveit config from my iron version of this
library (updated with a few jazzy changes).
When using the associated franka_description, it works in
simulation, even with updated libfranka.

The goal now is to see how we can transition this to the new franka_description.