Working on my own grippers
Trying to get mock grippers to work.
But the built-in franka gripper node is also running and interfering.

Forgot about that node.

Instead, should try to make that node work since it might behave more like
the real robot.

This commit will be in the history just in case, but I will revert it next
commit