Updated for changes to LoanedStateInterface API
It used to be possible to convert a StateInterface and CommandInterface to "Loaned" versions
directly. This possibility was removed in kilted, which broke tests: this is because
these "Loaned" interfaces are supposed to only come from the resource manager.

My understanding, however, is that for mocked/testing purposes, the interface not not be
managed by the resource manager.  The API still allows creating loaned interfaces from shared_ptr<StateInterface> so I create the state interface as shared pointers and then create LoanedInterfaces from them.

Whether this is the appropriate way of testing the behavior I don't know, but it does
make the code build and seems okay since this is just a mock/test.